# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BookingRequestInputPreferredContactMethod = typeof BookingRequestInputPreferredContactMethod[keyof typeof BookingRequestInputPreferredContactMethod];


export const BookingRequestInputPreferredContactMethod = {
  phone: 'phone',
  text: 'text',
  email: 'email',
} as const;

export type BookingRequestInputVehicleType = typeof BookingRequestInputVehicleType[keyof typeof BookingRequestInputVehicleType];


export const BookingRequestInputVehicleType = {
  sedan: 'sedan',
  coupe: 'coupe',
  suv: 'suv',
  truck: 'truck',
  van: 'van',
  other: 'other',
} as const;

/**
 * @nullable
 */
export type BookingRequestInputVehicleCondition = typeof BookingRequestInputVehicleCondition[keyof typeof BookingRequestInputVehicleCondition] | null;


export const BookingRequestInputVehicleCondition = {
  light: 'light',
  moderate: 'moderate',
  heavy: 'heavy',
  unsure: 'unsure',
} as const;

export interface BookingRequestInput {
  /** @minLength 2 */
  fullName: string;
  email: string;
  /** @minLength 7 */
  phone: string;
  preferredContactMethod: BookingRequestInputPreferredContactMethod;
  /** @minLength 1 */
  vehicleMake: string;
  /** @minLength 1 */
  vehicleModel: string;
  /** @nullable */
  vehicleYear?: number | null;
  vehicleType: BookingRequestInputVehicleType;
  /** @nullable */
  vehicleColor?: string | null;
  /** @nullable */
  vehicleCondition?: BookingRequestInputVehicleCondition;
  /** @minItems 1 */
  services: string[];
  /** @nullable */
  serviceNotes?: string | null;
  /** @minLength 3 */
  serviceAddress: string;
  /** @minLength 2 */
  city: string;
  /** @minLength 3 */
  postalCode: string;
  /** @nullable */
  locationType?: string | null;
  /** @nullable */
  accessNotes?: string | null;
  preferredDate: string;
  /** @minLength 1 */
  preferredTime: string;
  /** @nullable */
  alternativeDate?: string | null;
  /** @nullable */
  alternativeTime?: string | null;
  /** @nullable */
  notes?: string | null;
  consent: boolean;
  /** @nullable */
  website?: string | null;
}

export type BookingRequestStatus = typeof BookingRequestStatus[keyof typeof BookingRequestStatus];


export const BookingRequestStatus = {
  received: 'received',
  contacted: 'contacted',
  confirmed: 'confirmed',
  declined: 'declined',
} as const;

export type BookingRequest = BookingRequestInput & {
  id: number;
  submittedAt: string;
  status: BookingRequestStatus;
};

export interface ContactInquiryInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @minLength 2 */
  subject: string;
  /** @minLength 10 */
  message: string;
  /** @nullable */
  website?: string | null;
}

export type ContactInquiry = ContactInquiryInput & {
  id: number;
  submittedAt: string;
};

export interface ErrorResponse {
  error: string;
}