# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: bookings
    description: Appointment request operations
  - name: contact
    description: Contact inquiry operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /bookings:
    post:
      operationId: createBookingRequest
      tags: [bookings]
      summary: Submit an appointment request
      description: Creates a booking request for Wizard's Auto Detailing to review and confirm.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingRequestInput"
      responses:
        "201":
          description: Booking request received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingRequest"
        "400":
          description: Invalid booking request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /contact-inquiries:
    post:
      operationId: createContactInquiry
      tags: [contact]
      summary: Submit a contact inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInquiryInput"
      responses:
        "201":
          description: Contact inquiry received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactInquiry"
        "400":
          description: Invalid contact inquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    BookingRequestInput:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        phone:
          type: string
          minLength: 7
        preferredContactMethod:
          type: string
          enum: [phone, text, email]
        vehicleMake:
          type: string
          minLength: 1
        vehicleModel:
          type: string
          minLength: 1
        vehicleYear:
          type: ["integer", "null"]
        vehicleType:
          type: string
          enum: [sedan, coupe, suv, truck, van, other]
        vehicleColor:
          type: ["string", "null"]
        vehicleCondition:
          type: ["string", "null"]
          enum: [light, moderate, heavy, unsure, null]
        services:
          type: array
          minItems: 1
          items:
            type: string
        serviceNotes:
          type: ["string", "null"]
        serviceAddress:
          type: string
          minLength: 3
        city:
          type: string
          minLength: 2
        postalCode:
          type: string
          minLength: 3
        locationType:
          type: ["string", "null"]
        accessNotes:
          type: ["string", "null"]
        preferredDate:
          type: string
          format: date
        preferredTime:
          type: string
          minLength: 1
        alternativeDate:
          type: ["string", "null"]
          format: date
        alternativeTime:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        consent:
          type: boolean
        website:
          type: ["string", "null"]
      required:
        - fullName
        - email
        - phone
        - preferredContactMethod
        - vehicleMake
        - vehicleModel
        - vehicleType
        - services
        - serviceAddress
        - city
        - postalCode
        - preferredDate
        - preferredTime
        - consent
    BookingRequest:
      allOf:
        - $ref: "#/components/schemas/BookingRequestInput"
        - type: object
          properties:
            id:
              type: integer
            submittedAt:
              type: string
              format: date-time
            status:
              type: string
              enum: [received, contacted, confirmed, declined]
          required:
            - id
            - submittedAt
            - status
    ContactInquiryInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        phone:
          type: ["string", "null"]
        subject:
          type: string
          minLength: 2
        message:
          type: string
          minLength: 10
        website:
          type: ["string", "null"]
      required:
        - name
        - email
        - subject
        - message
    ContactInquiry:
      allOf:
        - $ref: "#/components/schemas/ContactInquiryInput"
        - type: object
          properties:
            id:
              type: integer
            submittedAt:
              type: string
              format: date-time
          required:
            - id
            - submittedAt
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error