# lib/api-zod/src/generated/types/bookingRequestInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BookingRequestInputPreferredContactMethod } from './bookingRequestInputPreferredContactMethod';
import type { BookingRequestInputVehicleCondition } from './bookingRequestInputVehicleCondition';
import type { BookingRequestInputVehicleType } from './bookingRequestInputVehicleType';

export interface BookingRequestInput {
  /** @minLength 2 */
  fullName: string;
  email: string;
  /** @minLength 7 */
  phone: string;
  preferredContactMethod: BookingRequestInputPreferredContactMethod;
  /** @minLength 1 */
  vehicleMake: string;
  /** @minLength 1 */
  vehicleModel: string;
  /** @nullable */
  vehicleYear?: number | null;
  vehicleType: BookingRequestInputVehicleType;
  /** @nullable */
  vehicleColor?: string | null;
  /** @nullable */
  vehicleCondition?: BookingRequestInputVehicleCondition;
  /** @minItems 1 */
  services: string[];
  /** @nullable */
  serviceNotes?: string | null;
  /** @minLength 3 */
  serviceAddress: string;
  /** @minLength 2 */
  city: string;
  /** @minLength 3 */
  postalCode: string;
  /** @nullable */
  locationType?: string | null;
  /** @nullable */
  accessNotes?: string | null;
  preferredDate: Date;
  /** @minLength 1 */
  preferredTime: string;
  /** @nullable */
  alternativeDate?: Date | null;
  /** @nullable */
  alternativeTime?: string | null;
  /** @nullable */
  notes?: string | null;
  consent: boolean;
  /** @nullable */
  website?: string | null;
}