# lib/api-zod/src/generated/types/contactInquiry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContactInquiryInput } from './contactInquiryInput';

export type ContactInquiry = ContactInquiryInput & {
  id: number;
  submittedAt: Date;
};