# lib/api-zod/src/generated/types/contactInquiryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInquiryInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @minLength 2 */
  subject: string;
  /** @minLength 10 */
  message: string;
  /** @nullable */
  website?: string | null;
}